requirements.txt
python-telegram-bot==20.3
pandas==2.2.2
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.1
openai
fpdf2>=2.6.1